Circle animation now changes angle based on candidate's poll numbers

+++ app/src/main/java/com/lumivote/lumivote/ui/candidate_tab/CircleAngleAnimation.java
@@ -10,7 +10,7 @@ public class CircleAngleAnimation extends Animation {
     private float oldAngle;
     private float newAngle;
 
-    public CircleAngleAnimation(Circle circle, int newAngle) {
+    public CircleAngleAnimation(Circle circle, float newAngle) {
         this.oldAngle = circle.getAngle();
         this.newAngle = newAngle;
         this.circle = circle;
